refactor(Notification): minor changes and renaming
