+++ src/components/Notification.legacy/_closeButton.scss
