+++ src/components/Notification.legacy/_notification.scss
@@ -1,0 +1,28 @@
+@import '../../globals/scss/vars';
+@import '../../globals/scss/colors';
+@import '../../globals/scss/typography';
+@import '../../globals/scss/spacing';
+@import '../../globals/scss/helper-mixins';
+@import '../../globals/scss/import-once';
+@import '../../globals/scss/css--reset';
+
+@import 'variables';
+@import 'toastContainer';
+@import 'toast';
+@import 'closeButton';
+@import 'progressBar';
+
+// entrance and exit animations
+@import 'animations/bounce.scss';
+//@import "animations/zoom.scss";
+//@import "animations/flip.scss";
+//@import "animations/slide.scss";
+
+@include exports('notification') {
+  /**
+ * Notifications (for react-toastify)
+ *
+ * Styling for all tables created
+ * with react-table
+ */
+}
