+++ src/components/Notification.legacy/_progressBar.scss
