+++ src/components/Notification.legacy/_variables.scss
