+++ src/components/Notification.legacy/animations/_bounce.scss
