+++ src/components/Notification.legacy/animations/_flip.scss
