+++ src/components/Notification.legacy/animations/_slide.scss
