+++ src/components/Notification/a11yIconWarningSolid.js
