+++ src/components/Notification/index.js
@@ -1,1 +1,1 @@
-export default from './Notification';
+export * from './Notification';
