--- src/components/NotificationNew/README.mdx
@@ -1,1 +1,0 @@
-### Usage
