--- src/components/NotificationNew/index.js
@@ -1,1 +1,0 @@
-export * from './Notification';
