+++ src/documentation/Icons/Icons.stories.mdx
@@ -1,0 +1,12 @@
+import { Meta, Story, Preview } from '@storybook/addon-docs/blocks';
+import Button from '../../components/Button';
+import Blockquote from '../../components/Blockquote';
+import Link from '../../components/Link';
+import IconsComponent from './IconsDocumentation';
+
+<Meta
+  title="Documentation/Icons"
+  subTitle="Icons & Logos for User Interfaces"
+/>
+
+<IconsComponent />
