+++ src/documentation/Icons/IconsDocumentation.js
@@ -121,106 +121,106 @@ from '@wfp/icons'`}
 
 const wrapperStyle = { display: 'flex', flexWrap: 'wrap', margin: '0 -0.7em' };
 
-storiesOf('Design|Core', module)
-  .addParameters({ options: { showPanel: false, isToolshown: false } })
-  .add('Icons', () => (
-    <Page title="Icons" subTitle="Icons & Logos for User Interfaces">
-      <p>
-        For Usage please take a look at the{' '}
-        <Link onClick={linkTo('Components|Icon', 'Default')}>Icon</Link>{' '}
-        component or{' '}
-        <Link onClick={linkTo(' Getting started', 'Developers')}>
-          Usage for Developers
-        </Link>
-        .
-      </p>
-
-      <Blockquote title="Notice">
-        WFP UI Icons is using a fork of{' '}
-        <Link href="https://github.com/IBM/carbon-icons" target="_blank">
-          Carbon Icons
-        </Link>
-        . You can find additional information in the{' '}
-        <Link
-          href="http://www.carbondesignsystem.com/guidelines/iconography"
-          target="_blank">
-          Carbon Design System
-        </Link>
-      </Blockquote>
-
-      <h3>Generate custom Icon</h3>
-      <p>
-        Generate a custom icon which can be used inside the{' '}
-        <span className="wfp--inline-highlight">icon</span> prop of the{' '}
-        <span className="wfp--inline-highlight">{`<Icon/>`}</span> component.
-        SVG images which are generated with Sketch App works best. Also consider
-        optimizing the svg image before.
-      </p>
-      <IconConverter />
-
-      <h3>Humanitarian Thematic Icon Set</h3>
-      <p>
-        The <i>UN Office for the Coordination of Humanitarian Affairs (OCHA)</i>{' '}
-        has established icons to represent humanitarian assistance and to ensure
-        consistency in use, these have been shared widely with UN agencies and
-        NGOs.{' '}
-        <Link
-          href="https://www.unocha.org/story/iconography-part-un%E2%80%99s-humanitarian-efforts-ocha-releases-new-humanitarian-icons"
-          target="_blank">
-          Visit OCHA Icon Library
-        </Link>
-      </p>
-      <p>
-        WFP has also created some{' '}
-        <Link
-          href="http://newgo.wfp.org/documents/humanitarian-thematic-icon-set-png"
-          target="_blank">
-          WFP specific icons
-        </Link>{' '}
-        to add to this set.
-      </p>
-      <h3>List of Icons</h3>
-
-      <Link
-        href="https://github.com/wfp/carbon-icons/tree/master/src/svg"
-        target="_blank">
-        Source svg files
+const IconsComponent = () => (
+  <>
+    <p>
+      For Usage please take a look at the{' '}
+      <Link onClick={linkTo('Components|Icon', 'Default')}>Icon</Link> component
+      or{' '}
+      <Link onClick={linkTo(' Getting started', 'Developers')}>
+        Usage for Developers
       </Link>
+      .
+    </p>
 
-      <h4>WFP Logos</h4>
-      <div style={wrapperStyle}>
-        {iconList
-          .filter(function(icon) {
-            return icon.name.includes('logo');
-          })
-          .map(icon => {
-            return <IconList icon={icon} />;
-          })}
-      </div>
+    <Blockquote title="Notice">
+      WFP UI Icons is using a fork of{' '}
+      <Link href="https://github.com/IBM/carbon-icons" target="_blank">
+        Carbon Icons
+      </Link>
+      . You can find additional information in the{' '}
+      <Link
+        href="http://www.carbondesignsystem.com/guidelines/iconography"
+        target="_blank">
+        Carbon Design System
+      </Link>
+    </Blockquote>
 
-      <h4>WFP Humanitatian Icons</h4>
-      <div style={wrapperStyle}>
-        {iconList
-          .filter(function(icon) {
-            return icon.name.includes('wfp-hum--');
-          })
-          .map(icon => {
-            return <IconList icon={icon} />;
-          })}
-      </div>
+    <h3>Generate custom Icon</h3>
+    <p>
+      Generate a custom icon which can be used inside the{' '}
+      <span className="wfp--inline-highlight">icon</span> prop of the{' '}
+      <span className="wfp--inline-highlight">{`<Icon/>`}</span> component. SVG
+      images which are generated with Sketch App works best. Also consider
+      optimizing the svg image before.
+    </p>
+    <IconConverter />
 
-      <h4>UI Icons</h4>
+    <h3>Humanitarian Thematic Icon Set</h3>
+    <p>
+      The <i>UN Office for the Coordination of Humanitarian Affairs (OCHA)</i>{' '}
+      has established icons to represent humanitarian assistance and to ensure
+      consistency in use, these have been shared widely with UN agencies and
+      NGOs.{' '}
+      <Link
+        href="https://www.unocha.org/story/iconography-part-un%E2%80%99s-humanitarian-efforts-ocha-releases-new-humanitarian-icons"
+        target="_blank">
+        Visit OCHA Icon Library
+      </Link>
+    </p>
+    <p>
+      WFP has also created some{' '}
+      <Link
+        href="http://newgo.wfp.org/documents/humanitarian-thematic-icon-set-png"
+        target="_blank">
+        WFP specific icons
+      </Link>{' '}
+      to add to this set.
+    </p>
+    <h3>List of Icons</h3>
 
-      <div style={wrapperStyle}>
-        {iconList
-          .filter(function(icon) {
-            return (
-              !icon.name.includes('wfp-hum--') && !icon.name.includes('logo')
-            );
-          })
-          .map(icon => {
-            return <IconList icon={icon} />;
-          })}
-      </div>
-    </Page>
-  ));
+    <Link
+      href="https://github.com/wfp/carbon-icons/tree/master/src/svg"
+      target="_blank">
+      Source svg files
+    </Link>
+
+    <h4>WFP Logos</h4>
+    <div style={wrapperStyle}>
+      {iconList
+        .filter(function (icon) {
+          return icon.name.includes('logo');
+        })
+        .map((icon) => {
+          return <IconList icon={icon} />;
+        })}
+    </div>
+
+    <h4>WFP Humanitatian Icons</h4>
+    <div style={wrapperStyle}>
+      {iconList
+        .filter(function (icon) {
+          return icon.name.includes('wfp-hum--');
+        })
+        .map((icon) => {
+          return <IconList icon={icon} />;
+        })}
+    </div>
+
+    <h4>UI Icons</h4>
+
+    <div style={wrapperStyle}>
+      {iconList
+        .filter(function (icon) {
+          return (
+            !icon.name.includes('wfp-hum--') && !icon.name.includes('logo')
+          );
+        })
+        .map((icon) => {
+          return <IconList icon={icon} />;
+        })}
+    </div>
+  </>
+);
+
+export default IconsComponent;
