+++ src/globals/scss/styles.scss
@@ -83,7 +83,7 @@ $deprecations--message: 'Deprecated code was found, this code will be removed be
 @import '../../components/Modal/modal';
 @import '../../components/Module/module';
 @import '../../components/Notification/notification';
-@import '../../components/NotificationNew/notification.scss';
+@import '../../components/Notification/notification.scss';
 @import '../../components/User/user';
 @import '../../components/Value/value';
 @import '../../components/InlineLoading/inline-loading';
